Have pe_metric_curl_cron_jobs send metrics to influxdb
instead of having telegraf collect them.

Prior to this commit, telegraf was setup on the dashboard node
and grabbed the metrics from various endpoints on the master
server.

After this commit, pe_metric_curl_cron_jobs is configured to
send metrics to influxdb in addition to storing them on disk.